♻️ refactor: 디바이스 관련 유틸 함수를 하나의 파일로 묶는다
