+++ src/utils/index.ts
@@ -1,5 +1,1 @@
-export * from './getOSByUserAgent';
-export * from './isAndroid';
-export * from './isIOS';
-export * from './isMobileWeb';
-export * from './isServer';
+export * from './device';
