--- src/utils/isAndroid.ts
@@ -1,9 +1,0 @@
-import { isServer } from './isServer';
-
-export const isAndroid = () => {
-  if (isServer()) {
-    return false;
-  }
-
-  return navigator.userAgent.match(/Android/i) !== null;
-};
