--- src/utils/isMobileWeb.ts
@@ -1,11 +1,0 @@
-/** @tossdocs-ignore */
-import { getOSByUserAgent } from './getOSByUserAgent';
-
-export const isMobileWeb = () => {
-  const userAgent = getOSByUserAgent();
-
-  if (userAgent === 'ios' || userAgent === 'android') {
-    return true;
-  }
-  return false;
-};
