--- src/utils/isServer.ts
@@ -1,3 +1,0 @@
-export const isServer = () => {
-  return typeof window === 'undefined' || 'Deno' in globalThis;
-};
